#13 Attempt a graceful recovery of the RCON connection when it's closed (TF2 closes/crashes).  Refactor bot detector log and command config code to new classes.
