+++ TF2SpectatorWin/TF2SpectatorWin.csproj
@@ -252,6 +252,7 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="ChatCommandDetails.cs" />
+    <Compile Include="CommandConfigModel.cs" />
     <Compile Include="Commands.xaml.cs">
       <DependentUpon>Commands.xaml</DependentUpon>
     </Compile>
@@ -259,6 +260,7 @@
     <Compile Include="Config.cs" />
     <Compile Include="CustomCommandFormat.cs" />
     <Compile Include="RelayCommand.cs" />
+    <Compile Include="TF2BotDetectorLogModel.cs" />
     <Compile Include="TF2SpectatorSettings.cs" />
     <Compile Include="TF2WindowsViewModel.cs" />
     <Compile Include="TwitchInstance.cs" />
